+++ DirectXTex/DirectXTexCompress.cpp
@@ -9,7 +9,7 @@
 // http://go.microsoft.com/fwlink/?LinkId=248926
 //-------------------------------------------------------------------------------------
 
-#include "DirectXTexp.h"
+#include "DirectXTexP.h"
 
 #ifdef _OPENMP
 #include <omp.h>
@@ -120,25 +120,25 @@ namespace
 
                 ptrdiff_t bytesLeft = pEnd - sptr;
                 assert(bytesLeft > 0);
-                size_t bytesToRead = std::min<size_t>(rowPitch, bytesLeft);
+                size_t bytesToRead = std::min<size_t>(rowPitch, static_cast<size_t>(bytesLeft));
                 if (!_LoadScanline(&temp[0], pw, sptr, bytesToRead, format))
                     return E_FAIL;
 
                 if (ph > 1)
                 {
-                    bytesToRead = std::min<size_t>(rowPitch, bytesLeft - rowPitch);
+                    bytesToRead = std::min<size_t>(rowPitch, static_cast<size_t>(bytesLeft - rowPitch));
                     if (!_LoadScanline(&temp[4], pw, sptr + rowPitch, bytesToRead, format))
                         return E_FAIL;
 
                     if (ph > 2)
                     {
-                        bytesToRead = std::min<size_t>(rowPitch, bytesLeft - rowPitch * 2);
+                        bytesToRead = std::min<size_t>(rowPitch, static_cast<size_t>(bytesLeft - rowPitch * 2));
                         if (!_LoadScanline(&temp[8], pw, sptr + rowPitch * 2, bytesToRead, format))
                             return E_FAIL;
 
                         if (ph > 3)
                         {
-                            bytesToRead = std::min<size_t>(rowPitch, bytesLeft - rowPitch * 3);
+                            bytesToRead = std::min<size_t>(rowPitch, static_cast<size_t>(bytesLeft - rowPitch * 3));
                             if (!_LoadScanline(&temp[12], pw, sptr + rowPitch * 3, bytesToRead, format))
                                 return E_FAIL;
                         }
